editing server set-up to use route and controller so that api queries are being handled in the api-controller. Next we'll want to test bringing the address from the front-end through to the api-controller to use that instead of the sample address we're currently using. Also added a catch-all route handler to give 404 errors, and a global error handler to catch any unknown middleware issues
