+++ server/routes/router.js
@@ -1,0 +1,15 @@
+const express = require('express');
+const controller = require('../controllers/api-controller.js');
+
+const router = express.Router();
+
+router.get(
+  '/',
+  controller.apiQueries,
+  (req, res) => {
+    console.log('saving final election data object for front end');
+    res.status(200).json(res.locals.actions);
+  },
+);
+
+module.exports = router;
